first version of search tool for db - command line

+++ max/src/orchidea_Max/main/orchidea.features.cpp
@@ -115,7 +115,8 @@ void* orchidea_dispatcher (void* d) {
     else {
         std::vector<float> features;
         try {
-            compute_features<float> (x->file_name->s_name, features, x->win_size, x->hop_size, 4, "moments");
+            compute_features<float> (x->file_name->s_name, features,
+                                     (int) x->win_size, x->hop_size, 4, "moments");
 
             //        compute_features<float> (x->file_name->s_name, features,
             //                                 x->win_size, x->hop_size, 4, "transients");
